Separate external and internal defs
